requirements.txt
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
idna==2.10
iniconfig==1.1.1
numpy==1.20.3
onnx==1.9.0
onnxruntime==1.7.0
packaging==21.0
Pillow==8.2.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-timeout==1.4.2
pyzmp==0.0.17
pyzmq==22.2.1
requests==2.25.1
scipy==1.7.1
SCons==4.2.0
six==1.16.0
tblib==1.7.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.5
